+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `lvl update` allows you to easily update `lvl` to the latest version.
 * Fix API marshalling errors with system checks and app component types.
 * You can now refer to system checks by type in commands, e.g. `system check delete my.cool.system disk`
+* You can now enable or disable DKIM on mail domains with `lvl mail domain dkim`.
 
 ## 1.6.0